# apps/ui/src/api/openapi.json
{
  "components": {
    "schemas": {
      "ClaudeNote": {
        "additionalProperties": false,
        "description": "Structured qualitative note authored by Claude.\n\nFile-based ingest: callers write ``<notes_dir>/match-<id>.json`` and the\nAPI parses it via :meth:`model_validate_json` on read.",
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "match_id": {
            "title": "Match Id",
            "type": "integer"
          },
          "qualitative_deltas": {
            "items": {
              "discriminator": {
                "mapping": {
                  "1x2": "#/components/schemas/Delta1x2",
                  "btts": "#/components/schemas/DeltaBTTS",
                  "corners_total": "#/components/schemas/DeltaCornersTotal",
                  "ou_2_5": "#/components/schemas/DeltaOU25"
                },
                "propertyName": "market"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Delta1x2"
                },
                {
                  "$ref": "#/components/schemas/DeltaOU25"
                },
                {
                  "$ref": "#/components/schemas/DeltaBTTS"
                },
                {
                  "$ref": "#/components/schemas/DeltaCornersTotal"
                }
              ]
            },
            "title": "Qualitative Deltas",
            "type": "array"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "title": "Sources",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "match_id",
          "created_at",
          "summary",
          "qualitative_deltas",
          "confidence",
          "sources"
        ],
        "title": "ClaudeNote",
        "type": "object"
      },
      "Delta1x2": {
        "properties": {
          "log_odds_shift": {
            "title": "Log Odds Shift",
            "type": "number"
          },
          "market": {
            "const": "1x2",
            "default": "1x2",
            "title": "Market",
            "type": "string"
          }
        },
        "required": [
          "log_odds_shift"
        ],
        "title": "Delta1x2",
        "type": "object"
      },
      "DeltaBTTS": {
        "properties": {
          "log_odds_shift": {
            "title": "Log Odds Shift",
            "type": "number"
          },
          "market": {
            "const": "btts",
            "default": "btts",
            "title": "Market",
            "type": "string"
          }
        },
        "required": [
          "log_odds_shift"
        ],
        "title": "DeltaBTTS",
        "type": "object"
      },
      "DeltaCornersTotal": {
        "properties": {
          "log_odds_shift": {
            "title": "Log Odds Shift",
            "type": "number"
          },
          "market": {
            "const": "corners_total",
            "default": "corners_total",
            "title": "Market",
            "type": "string"
          }
        },
        "required": [
          "log_odds_shift"
        ],
        "title": "DeltaCornersTotal",
        "type": "object"
      },
      "DeltaOU25": {
        "properties": {
          "log_odds_shift": {
            "title": "Log Odds Shift",
            "type": "number"
          },
          "market": {
            "const": "ou_2_5",
            "default": "ou_2_5",
            "title": "Market",
            "type": "string"
          }
        },
        "required": [
          "log_odds_shift"
        ],
        "title": "DeltaOU25",
        "type": "object"
      },
      "Fixture": {
        "description": "Compact match row for ``GET /fixtures``.",
        "properties": {
          "away_team": {
            "title": "Away Team",
            "type": "string"
          },
          "competition": {
            "title": "Competition",
            "type": "string"
          },
          "home_team": {
            "title": "Home Team",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kickoff_utc": {
            "format": "date-time",
            "title": "Kickoff Utc",
            "type": "string"
          },
          "season": {
            "title": "Season",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "competition",
          "season",
          "home_team",
          "away_team",
          "kickoff_utc",
          "status"
        ],
        "title": "Fixture",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MatchDetail": {
        "description": "``GET /matches/{id}`` payload \u2014 adds final score when known.",
        "properties": {
          "away_goals": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Away Goals"
          },
          "away_team": {
            "title": "Away Team",
            "type": "string"
          },
          "competition": {
            "title": "Competition",
            "type": "string"
          },
          "home_goals": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Goals"
          },
          "home_team": {
            "title": "Home Team",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kickoff_utc": {
            "format": "date-time",
            "title": "Kickoff Utc",
            "type": "string"
          },
          "season": {
            "title": "Season",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "competition",
          "season",
          "home_team",
          "away_team",
          "kickoff_utc",
          "status"
        ],
        "title": "MatchDetail",
        "type": "object"
      },
      "PredictCachedResponse": {
        "description": "200 response \u2014 pre-existing prediction returned without re-fitting.",
        "properties": {
          "cached": {
            "const": true,
            "default": true,
            "title": "Cached",
            "type": "boolean"
          },
          "markets": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "number"
              },
              "type": "object"
            },
            "title": "Markets",
            "type": "object"
          },
          "match_id": {
            "title": "Match Id",
            "type": "integer"
          },
          "model_run_id": {
            "title": "Model Run Id",
            "type": "integer"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          }
        },
        "required": [
          "match_id",
          "model_version",
          "model_run_id",
          "markets"
        ],
        "title": "PredictCachedResponse",
        "type": "object"
      },
      "PredictEnqueuedResponse": {
        "description": "202 response \u2014 fit enqueued, follow up by ``model_run_id``.",
        "properties": {
          "cached": {
            "const": false,
            "default": false,
            "title": "Cached",
            "type": "boolean"
          },
          "model_run_id": {
            "title": "Model Run Id",
            "type": "integer"
          },
          "status": {
            "const": "running",
            "default": "running",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "model_run_id"
        ],
        "title": "PredictEnqueuedResponse",
        "type": "object"
      },
      "PredictRequest": {
        "description": "Body for ``POST /matches/{id}/predict``.",
        "properties": {
          "force_refit": {
            "default": false,
            "title": "Force Refit",
            "type": "boolean"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          }
        },
        "title": "PredictRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Phase 0 football match outcome predictor.",
    "title": "football-predictor",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/fixtures": {
      "get": {
        "description": "Return all scheduled matches sorted by ``kickoff_utc`` ascending.",
        "operationId": "list_fixtures_fixtures_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Fixture"
                  },
                  "title": "Response List Fixtures Fixtures Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Fixtures"
      }
    },
    "/matches/{match_id}": {
      "get": {
        "operationId": "get_match_matches__match_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Match"
      }
    },
    "/matches/{match_id}/notes": {
      "get": {
        "operationId": "get_match_note_matches__match_id__notes_get",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaudeNote"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Match Note"
      }
    },
    "/matches/{match_id}/predict": {
      "post": {
        "operationId": "predict_matches__match_id__predict_post",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "title": "Match Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictCachedResponse",
                  "title": "Response Predict Matches  Match Id  Predict Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictEnqueuedResponse"
                }
              }
            },
            "description": "Accepted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Predict"
      }
    }
  }
}